refactor of pages placement in the file structure
